requirements.txt
apache-airflow==2.9.3
pandas>=2.2.2
yfinance>=0.2.50
SQLAlchemy>=2.0.30
python-dotenv>=1.0.1
streamlit>=1.36.0
plotly>=5.22.0



